apps/mint/pyproject.toml
[project]
name = "mint-core"
version = "0.1.0"
description = "Mint - Classic Maplestory community hub"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "django>=5.2.7",
  "django-vite>=3.1.0",
  "djangorestframework>=3.16.1",
  "drf-spectacular>=0.28.0",
  "psycopg2-binary>=2.9.11",
  "python-dotenv>=1.1.1",
  "django-simple-history>=3.7.0",
  "drf-flex-fields>=1.0.2",
  "django-filter>=25.2",
]

[project.optional-dependencies]
dev = [
  "django-stubs[compatible-mypy]>=5.2.7",
  "djangorestframework-stubs[compatible-mypy]>=3.16.4",
  "black>=24.0.0",
  "isort>=5.13.0",
]

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
